+++ src/engine/node-html-parser.ts
@@ -1,22 +1,44 @@
-import { Node, SearchContext, SearchEngine } from "./search-engine";
-import * as HTMLParser from "node-html-parser";
 import * as CSSselect from "css-select";
+import { pHtmlParser } from "html-parser";
+import { IPHtmlDocument, IPHtmlElement, IPHtmlNode } from "html-parser/dist/interface";
+import { PHtmlDocument } from "html-parser/dist/model/PHtmlDocument";
+import { PHtmlElement } from "html-parser/dist/model/PHtmlElement";
+import { PHtmlNode } from "html-parser/dist/model/PHtmlNode";
+import * as vscode from "vscode";
+import { htmlUtil } from "../util/html-util";
+import { QSNode, SearchContext } from "../model/search-context.model";
+import { SearchEngine } from "./search-engine";
 
 export class NodeHtmlParserAdaptor extends SearchEngine {
-  validateSearchContext(queryExpr: SearchContext){
-    const compiledQuery = CSSselect.compile(queryExpr.search);
+  canApply( uri: vscode.Uri){
+    return uri.path.endsWith(".html") || uri.path.endsWith(".htm");
+  }
+  createClone(v: QSNode): QSNode {
+    if( v instanceof PHtmlElement || v instanceof PHtmlNode || v instanceof PHtmlDocument ){
+      return v.cloneNode()!;
+    }
+    throw new Error("unexpected error");
+  }
+
+   getRange(baseTag: QSNode): readonly [number, number] {
+    const range = htmlUtil.getOffsetOfCloseTag(baseTag);
+    return [range.startOffset, range.endOffset];
+  }
+
+
+  parser = new pHtmlParser({skipComment: false});
+  validateSearchContext(searchContext: SearchContext){
+    const compiledQuery = CSSselect.compile(searchContext.search);
 
     return true;
   };
-  searchHtml(content: string, queryExpr: SearchContext): Node[] {
-    const rootNode = HTMLParser.parse(content, {
-      comment: true,
-      voidTag: { closingSlash: false },
-    });
-    const result = rootNode.querySelectorAll(queryExpr.search);
-    return result;
+  searchHtml(content: string, searchContext: SearchContext): QSNode[] {
+    const rootNode = this.parser.parse(content);
+    const result = rootNode.querySelectorAll(searchContext.search);
+    return [...result];
   }
-  getReplacedText(node: Node): string{
-    return node.parentNode?.toString();
+  getReplacedText(node: QSNode): string{
+    return node.parentNode?.outerHTML ?? "";
   };
+  
 }