+++ src/engine/replace-edit.ts
@@ -1,32 +1,31 @@
-import { Node } from "node-html-parser";
-import { TextDecoder } from "util";
+import * as  l10n from "@vscode/l10n";
 import * as vscode from "vscode";
 import { Range } from "vscode";
-
-export type ReplaceDocument =
-  | vscode.TextDocument
-  | { uri: vscode.Uri; getText(range?: Range): string };
 
 type ReplaceOffset = readonly [number, number];
 
 interface EditRecord {}
 
 export interface ReplaceEdit {
-  openTextDocument(uri: vscode.Uri): Thenable<ReplaceDocument>;
+  openTextDocument(uri: vscode.Uri): Thenable<vscode.TextDocument>;
   replace(
     uri: vscode.Uri,
     range: ReplaceOffset,
     newText: string
   ): Thenable<void>;
   applyEdit(): Thenable<void>;
-  modifiedTextDocument(uri: vscode.Uri): Thenable<ReplaceDocument>;
+  modifiedTextDocument(uri: vscode.Uri): Thenable<vscode.TextDocument>;
 }
 
 export class ReplaceEditTextDocument implements ReplaceEdit {
-  private edit = new vscode.WorkspaceEdit();
+  protected edit = new vscode.WorkspaceEdit();
+  document!: vscode.TextDocument;
 
   async openTextDocument(uri: vscode.Uri): Promise<vscode.TextDocument> {
-    return vscode.workspace.openTextDocument(uri!);
+    if(!this.document){
+      this.document = await vscode.workspace.openTextDocument(uri!);
+    }
+    return this.document;
   }
   async replace(
     uri: vscode.Uri,
@@ -41,43 +40,50 @@ export class ReplaceEditTextDocument implements ReplaceEdit {
     this.edit.replace(uri, vsrange, newText);
   }
   async applyEdit(): Promise<void> {
-    await vscode.workspace.applyEdit(this.edit);
+    if(this.edit.size === 0){
+      return;
+    }
+    const result = await vscode.workspace.applyEdit(this.edit);
+    if( !result ){
+      throw new Error(vscode.l10n.t("cannot applyEdit."));
+    } 
+    await this.document.save();
   }
 
-  async modifiedTextDocument(uri: vscode.Uri): Promise<ReplaceDocument> {
-    return vscode.workspace.openTextDocument(uri!);
+  async modifiedTextDocument(uri: vscode.Uri): Promise<vscode.TextDocument> {
+    return this.document;
   }
 }
 
-export class ReplaceEditInMemory implements ReplaceEdit {
-  cache: Map<string, string> = new Map();
-  offset: Map<string, number> = new Map();
+export class ReplaceEditMemFsTextDocument extends ReplaceEditTextDocument {
+  document!: vscode.TextDocument;
 
-  async openTextDocument(uri: vscode.Uri): Promise<ReplaceDocument> {
-    const content = (await vscode.workspace.openTextDocument(uri!)).getText();
-    
-    this.cache.set(uri.path, content);
-    this.offset.set(uri.path, 0);
-    return { uri: uri, getText: () => this.cache.get(uri.path)! };
-  }
-  async replace(
-    uri: vscode.Uri,
-    range: ReplaceOffset,
-    newText: string
-  ): Promise<void> {
-    let content = this.cache.get(uri.path)!;
-    let offset = this.offset.get(uri.path)!;
-    content =
-      content.substring(0, range[0] - offset) +
-      newText +
-      content.substring(range[1] - offset);
-    offset += range[1] - range[0] - newText.length;
-    this.cache.set(uri.path, content);
-    this.offset.set(uri.path, offset);
-  }
-  async applyEdit(): Promise<void> {}
+  async openTextDocument(uri: vscode.Uri): Promise<vscode.TextDocument> {
+    if(!this.document){
+      const memUri = vscode.Uri.from({scheme: "memfs", path: uri.path, fragment: uri.scheme });
+      const document = await vscode.workspace.openTextDocument(memUri);
 
-  async modifiedTextDocument(uri: vscode.Uri): Promise<ReplaceDocument> {
-    return { uri: uri, getText: () => this.cache.get(uri.path)! };
+      // XXX workaround reload
+      const edit = new vscode.WorkspaceEdit();
+      const range = new vscode.Range(
+        document.lineAt(0).range.start,
+        document.lineAt(document.lineCount - 1).range.end
+      );
+      const rawText = await vscode.workspace.fs.readFile(vscode.Uri.from({scheme: uri.fragment, path: uri.path}));
+      edit.replace(document.uri, range, rawText.toString());
+      const result = await vscode.workspace.applyEdit(edit);
+      this.document = document;
+    }
+    return this.document;
+  }
+
+  async applyEdit(): Promise<void> {
+    if(this.edit.size === 0){
+      return;
+    }
+    const result = await vscode.workspace.applyEdit(this.edit);
+    if( !result ){
+      throw new Error(vscode.l10n.t("cannot applyEdit."));
+    } 
   }
 }