+++ src/extension.ts
@@ -3,14 +3,18 @@ import { SearchQueryPanelProvider } from "./view/search-query-panel";
 import { SearchResultPanelProvider } from "./view/search-result-panel";
 import { ReplacePreviewDocumentProvider } from "./view/replace-preview";
 import { NodeHtmlParserAdaptor } from "./engine/node-html-parser";
+import { JsxHtmlParserAdapter } from "./engine/jsx-htmlnode-parser";
 
 export function activate(context: vscode.ExtensionContext) {
-  const searchEngine = new NodeHtmlParserAdaptor();
-  const result = new SearchResultPanelProvider(searchEngine);
-  result.init(context);
-  new SearchQueryPanelProvider(context, result);
-  new ReplacePreviewDocumentProvider(searchEngine).init(context);
-  
+  const searchEngines = [
+    new NodeHtmlParserAdaptor(),
+    new JsxHtmlParserAdapter(),
+  ];
+  const previewProvider = new ReplacePreviewDocumentProvider(...searchEngines);
+  const resultPanel = new SearchResultPanelProvider(previewProvider, ...searchEngines);
+  previewProvider.init(context);
+  resultPanel.init(context);
+  new SearchQueryPanelProvider(context, resultPanel, previewProvider);
 }
 
 export function deactivate() {}