+++ src/test/suite/extension.test.ts
@@ -5,7 +5,7 @@ import * as vscode from "vscode";
 
 suite("Extension Test Suite", () => {
   after(() => {
-    vscode.window.showInformationMessage("All tests done!");
+    void vscode.window.showInformationMessage("All tests done!");
   });
 
   suiteSetup(async () => {});