+++ src/util/html-util.ts
@@ -1,22 +1,12 @@
-import * as HTMLParser from "node-html-parser";
-import { Node } from "../engine/search-engine";
+import { PHtmlElement,PHtmlNode,PHtmlDocument } from 'html-parser';
+import { QSNode } from "../model/search-context.model";
 
 export namespace htmlUtil {
-  export function getOffsetOfOpenTag(v: HTMLParser.HTMLElement) {
-    const startOffset = v.range[0];
-    const endOffset = Math.min(
-      ...[v.range[1], v.firstChild?.range[0], v.nextSibling?.range[0]].filter(
-        (v) => v
-      )
-    );
-
-    return { startOffset, endOffset };
-  }
-
-  export function getOffsetOfCloseTag(v: Node) {
-    const startOffset = v.range[0];
-    const endOffset = v.range[1];
-
-    return { startOffset, endOffset };
+  export function getOffsetOfCloseTag(v: QSNode) {
+    if( v instanceof PHtmlElement || v instanceof PHtmlNode || v instanceof PHtmlDocument ){
+      const r = v.range!;
+      return { startOffset:r.startOpenTag, endOffset:r.endCloseTag };
+    }
+    return { startOffset:0, endOffset:0 };
   }
 }