+++ src/view/replace-preview.ts
@@ -6,23 +6,29 @@ import {
   TextDocumentContentProvider,
   Uri,
 } from "vscode";
-import { Constants } from "../constants";
-import { SearchResultPanelProvider } from "./search-result-panel";
-import { SearchContext, SearchEngine } from "../engine/search-engine";
-import { ReplaceEditInMemory } from "../engine/replace-edit";
 
+import { Constants } from "../constants";
+import { SearchEngine } from "../engine/search-engine";
+import { MemFS } from "../engine/inmemory-fsprovider";
+import { ReplaceEditMemFsTextDocument } from "../engine/replace-edit";
+import { SearchContext } from "../model/search-context.model";
+import { SerachResultItem } from "../model/search-result.model";
+import * as path from 'path';
 export class ReplacePreviewDocumentProvider
   implements TextDocumentContentProvider, Disposable
 {
-  searchContext: SearchContext = <any>{};
-  constructor(private searchEngine: SearchEngine) {}
+  searchEngines: SearchEngine[];
+  constructor(..._searchEngines: SearchEngine[]) {
+    this.searchEngines = _searchEngines;
+  }
 
   init(context: vscode.ExtensionContext) {
     context.subscriptions.push(
       vscode.workspace.registerTextDocumentContentProvider(
         Constants.SCHEMA_PREVIEW,
         this
-      )
+      ),
+      vscode.workspace.registerFileSystemProvider("memfs", new MemFS, {isReadonly: true})
     );
   }
 
@@ -35,6 +41,8 @@ export class ReplacePreviewDocumentProvider
     this._onDidChange.dispose();
   }
 
+  lock: Record<string, number> = {};
+  lockNo = 0;
   async provideTextDocumentContent(
     uri: vscode.Uri,
     token: vscode.CancellationToken
@@ -43,17 +51,74 @@ export class ReplacePreviewDocumentProvider
       return "Canceled";
     }
 
-    const edit = new ReplaceEditInMemory();
-    const text = await edit.openTextDocument(
-      uri.with({ scheme: uri.fragment, path: uri.path })
-    );
-    const searchContext = JSON.parse(uri.query);
-    const searchResult = await this.searchEngine.search(text, searchContext);
-    if (!searchResult){
+    try{
+      const edit = new ReplaceEditMemFsTextDocument();
+      const text = await edit.openTextDocument(
+        Uri.from({ scheme: uri.fragment, path: uri.path })
+      );
+      const {searchContext} = JSON.parse(uri.query) ;
+      const searchEngine = this.searchEngines.find(v=>v.canApply(uri));
+      if( !searchEngine){
+        console.warn("can not found appliable engine.");
+        return "";
+      }
+  
+      const searchResult = searchEngine.search(text, searchContext);
+      if (!searchResult){
+        return "";
+      }
+      // searchResult.items = searchResult.items.filter(v=>v.index === index);
+      try{
+        await searchEngine.replace(searchResult, searchContext.replace, edit);
+      }catch(e){
+        console.log(e);
+      }
+      return text.getText();
+    }catch(e){
       return "";
     }
-    await this.searchEngine.replace([searchResult], searchContext.replace, edit);
-
-    return text.getText();
   }
+
+  async openReplacePreview(item: SerachResultItem, searchContext: SearchContext) {
+    const start = item.startOffset;
+    const end = item.endOffset;
+    const preview = getPreviewUri(item.document.uri, searchContext, item);
+    const document = item.document;
+    {
+      const range = new vscode.Range(
+        document.positionAt(start),
+        document.positionAt(end)
+      );
+      const options: vscode.TextDocumentShowOptions = {
+        selection: range,
+        preserveFocus: true,
+        preview: true,
+      };
+      const basefile = path.basename(item.document.uri.fsPath);
+
+      await vscode.commands.executeCommand(
+        "vscode.diff",
+        item.document.uri,
+        preview,
+        vscode.l10n.t(`{0} ↔ [replaced] (Replace Preview)`, basefile),
+        options
+      );
+      // this._onDidChange.fire(preview);
+    };
+  }
+
+  refresh(resourceUri: vscode.Uri, searchContext: SearchContext) {
+    this._onDidChange.fire(getPreviewUri(resourceUri, searchContext));
+  }
+  
 }
+
+function getPreviewUri(resource: vscode.Uri, searchContext: SearchContext, item?: SerachResultItem) {
+  return vscode.Uri.from({
+    scheme: Constants.SCHEMA_PREVIEW,
+    path: resource.path,
+    fragment: resource.scheme,
+    query: JSON.stringify({ searchContext }),
+  });
+}
+