+++ webpack.config.js
@@ -21,6 +21,7 @@ const extensionConfig = {
   },
   externals: {
     vscode: 'commonjs vscode' // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
+    ,"../html-parser": 'commonjs html-parser'
     // modules added here also need to be added in the .vscodeignore file
   },
   resolve: {